requirements.txt
# LangChain Core
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-community

# LangGraph (핵심)
langgraph>=0.0.40

# Web Search Tools
tavily-python>=0.3.0

# Web Scraping
beautifulsoup4>=4.12.0
httpx>=0.25.0
aiohttp>=3.9.0

# Web UI
streamlit>=1.29.0

# Environment & Utils
python-dotenv>=1.0.0
pydantic>=2.0.0

# Markdown & Report
markdown>=3.5.0
rich>=13.0.0

# Optional: Alternative search
# google-serper>=0.0.6
